меняет структуру редьюсеров, добавляет reselect
