--- src/consts.js
@@ -1,8 +1,0 @@
-export const reduxActionTypes = {
-  REQUIRED_AUTHORIZATION: `REQUIRED_AUTHORIZATION`,
-
-  LOAD_MOVIES: `LOAD_MOVIES`,
-  SET_FILTER_CATEGORY: `SET_FILTER_CATEGORY`,
-  GET_FILTERED_MOVIES: `GET_FILTERED_MOVIES`,
-  GET_GENRES_LIST: `GET_GENRES_LIST`,
-};
