+++ src/reducer/movies/actionTypes.js
@@ -1,0 +1,6 @@
+const actionTypes = {
+  LOAD_MOVIES: `LOAD_MOVIES`,
+  SET_FILTER_CATEGORY: `SET_FILTER_CATEGORY`,
+};
+
+export default actionTypes;
