+++ src/reducer/movies/actions.js
@@ -1,12 +1,9 @@
-import {reduxActionTypes} from './consts';
+import actionTypes from './actionTypes';
 
 const {
   SET_FILTER_CATEGORY,
-  GET_FILTERED_MOVIES,
-  GET_GENRES_LIST,
-  REQUIRED_AUTHORIZATION,
   LOAD_MOVIES,
-} = reduxActionTypes;
+} = actionTypes;
 
 export const ActionCreator = {
   loadMovies: (data) => ({
@@ -18,19 +15,6 @@ export const ActionCreator = {
     type: SET_FILTER_CATEGORY,
     payload: categoryName,
   }),
-
-  getFilteredMovies: () => ({
-    type: GET_FILTERED_MOVIES
-  }),
-
-  getGenresList: () => ({
-    type: GET_GENRES_LIST
-  }),
-
-  requireAuthorization: (status) => ({
-    type: REQUIRED_AUTHORIZATION,
-    payload: status,
-  }),
 };
 
 export const Operation = {
@@ -38,8 +22,6 @@ export const Operation = {
     return api.get(`/films`)
       .then((response) => {
         dispatch(ActionCreator.loadMovies(response.data));
-        dispatch(ActionCreator.getFilteredMovies());
-        dispatch(ActionCreator.getGenresList());
       });
   },
 };
