+++ src/reducer/name-spaces.js
@@ -1,0 +1,6 @@
+const NameSpace = {
+  MOVIES: `MOVIES`,
+  USER: `USER`
+};
+
+export default NameSpace;
