+++ src/reducer/user/actionTypes.js
@@ -1,0 +1,5 @@
+const actionTypes = {
+  REQUIRED_AUTHORIZATION: `REQUIRED_AUTHORIZATION`,
+};
+
+export default actionTypes;
